Fix BtlUnit_GetHeight return type s16->s32, implement accrobat_timing_icon_disp
- BtlUnit_GetHeight: returns s32 to avoid extra extsh in callers
- accrobat_timing_icon_disp: full implementation (pos+15, height+5, iconDispGx)
- battle_acrobat data: 24/32 -> 32/32 fully matched

+++ include/battle/battle_unit.h
@@ -620,7 +620,7 @@ void BtlUnit_SetPartsBaseScale(BattleWorkUnitPart* part, f32 x, f32 y, f32 z);
 void BtlUnit_SetPartsScale(BattleWorkUnitPart* part, f32 x, f32 y, f32 z);
 void BtlUnit_AddPartsScale(BattleWorkUnitPart* part, f32 x, f32 y, f32 z);
 s16 BtlUnit_GetWidth(BattleWorkUnit* unit);
-s16 BtlUnit_GetHeight(BattleWorkUnit* unit);
+s32 BtlUnit_GetHeight(BattleWorkUnit* unit);
 void BtlUnit_SetHeight(BattleWorkUnit* unit, s16 height);
 void BtlUnit_SetOffsetPos(BattleWorkUnit* unit, f32 x, f32 y, f32 z);
 void BtlUnit_GetPartsOffsetPos(BattleWorkUnitPart* part, f32* x, f32* y, f32* z);
